Makefile refinements & Vue App Entrypoint
Make some changes to the Makefile to refine it and add a watch target
that uses `reflex` to watch files and rebuild the vue app and reload the
go server.

Also made some changes to the Go app to allow Echo to serve the static
files and added the entrypoint for the vue app.